with_timer_in_different_thread_with_timeout_returns_false_after_timeout_message_was_sent
