+++ lumen_runtime/src/otp/erlang/read_timer_2/test/with_reference/with_list_options/without_async/with_timer/in_different_thread.rs
@@ -40,23 +40,5 @@ fn without_timeout_returns_milliseconds_remaining() {
 
 #[test]
 fn with_timeout_returns_false_after_timeout_message_was_sent() {
-    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
-        timeout_after_half_and_wait(milliseconds, barrier);
-        timeout_after_half_and_wait(milliseconds, barrier);
-
-        let timeout_message = different_timeout_message(timer_reference, process);
-
-        assert_has_message!(process, timeout_message);
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(false.into())
-        );
-
-        // again
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(false.into())
-        );
-    });
+    crate::test::with_timer_in_different_thread_with_timeout_returns_false_after_timeout_message_was_sent(native, options);
 }
